fix the 'highest score' not being the max score after winning the game
the score and highest score would not be updated to the max score due to
a re-render being triggered on winning the game setting the score to 0
before it is updated to the max score.
before the win condition was checked with the `seen` arrays length, now
it is checked by score.